# docs/openapi.json
{
  "components": {
    "schemas": {
      "Assignment": {
        "properties": {
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "title": "Mode",
            "type": "string"
          },
          "question_set_id": {
            "format": "uuid",
            "title": "Question Set Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus",
            "default": "assigned"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "family_id",
          "question_set_id",
          "child_id"
        ],
        "title": "Assignment",
        "type": "object"
      },
      "AssignmentStatus": {
        "enum": [
          "assigned",
          "in_progress",
          "submitted",
          "grading",
          "results_ready",
          "correcting",
          "completed",
          "withdrawn",
          "stopped"
        ],
        "title": "AssignmentStatus",
        "type": "string"
      },
      "AssignmentWork": {
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/Assignment"
          },
          "attempt": {
            "$ref": "#/components/schemas/Attempt"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "assignment",
          "attempt",
          "questions"
        ],
        "title": "AssignmentWork",
        "type": "object"
      },
      "Attempt": {
        "properties": {
          "assignment_id": {
            "format": "uuid",
            "title": "Assignment Id",
            "type": "string"
          },
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sequence": {
            "default": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "required": [
          "family_id",
          "assignment_id",
          "child_id"
        ],
        "title": "Attempt",
        "type": "object"
      },
      "AttemptResults": {
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QuestionResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "attempt_id",
          "complete",
          "results"
        ],
        "title": "AttemptResults",
        "type": "object"
      },
...
      "ChildAssignmentSummary": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_attempt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Attempt Id"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "question_count": {
            "title": "Question Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatus"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "mode",
          "time_limit_seconds",
          "question_count"
        ],
        "title": "ChildAssignmentSummary",
        "type": "object"
      },
...
      "CreateAssignmentRequest": {
        "properties": {
          "child_id": {
            "format": "uuid",
            "title": "Child Id",
            "type": "string"
          },
          "mode": {
            "default": "practice",
            "enum": [
              "practice",
              "exam"
            ],
            "title": "Mode",
            "type": "string"
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "maximum": 14400.0,
                "minimum": 60.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds"
          }
        },
        "required": [
          "child_id"
        ],
        "title": "CreateAssignmentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrintableAssignment": {
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/Assignment"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "title": "Questions",
            "type": "array"
          },
          "template_version": {
            "const": "a4-v1",
            "default": "a4-v1",
            "title": "Template Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "assignment",
          "title",
          "questions"
        ],
        "title": "PrintableAssignment",
        "type": "object"
      },
...
      "QuestionResult": {
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "awarded_points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awarded Points"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "feedback": {
            "additionalProperties": true,
            "title": "Feedback",
            "type": "object"
          },
          "grader_version": {
            "default": "fixture-v1",
            "title": "Grader Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/GradingOutcome"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "attempt_id",
          "question_id",
          "outcome",
          "awarded_points",
          "confidence",
          "feedback"
        ],
        "title": "QuestionResult",
        "type": "object"
      },
...
      "QuestionType": {
        "enum": [
          "single_choice",
          "multiple_choice",
          "typed_text",
          "word_order",
          "handwriting",
          "photo",
          "listening"
        ],
        "title": "QuestionType",
        "type": "string"
      },
      "QuestionView": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "required": [
          "id",
          "position",
          "type",
          "prompt",
          "points"
        ],
        "title": "QuestionView",
        "type": "object"
      },
      "ResponseKind": {
        "enum": [
          "choice",
          "text",
          "tokens",
          "strokes",
          "photo"
        ],
        "title": "ResponseKind",
        "type": "string"
      },
...
      "SaveResponseRequest": {
        "properties": {
          "answer": {
            "additionalProperties": true,
            "title": "Answer",
            "type": "object"
          },
          "expected_version": {
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/ResponseKind"
          }
        },
        "required": [
          "kind",
          "answer",
          "expected_version"
        ],
        "title": "SaveResponseRequest",
        "type": "object"
      },
      "SavedResponse": {
        "properties": {
          "answer": {
            "additionalProperties": true,
            "title": "Answer",
            "type": "object"
          },
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "family_id": {
            "format": "uuid",
            "title": "Family Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ResponseKind"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "family_id",
          "attempt_id",
          "question_id",
          "kind",
          "answer"
        ],
        "title": "SavedResponse",
        "type": "object"
      },
      "SubmissionReceipt": {
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/Assignment"
          },
          "attempt": {
            "$ref": "#/components/schemas/Attempt"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "required": [
          "assignment",
          "attempt",
          "job"
        ],
        "title": "SubmissionReceipt",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Learning Assessment API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/assignments": {
      "get": {
        "operationId": "list_child_assignments_v1_assignments_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChildAssignmentSummary"
                  },
                  "title": "Response List Child Assignments V1 Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Child Assignments",
        "tags": [
          "assignments"
        ]
      }
    },
    "/v1/assignments/{assignment_id}/printable": {
      "get": {
        "operationId": "get_printable_assignment_v1_assignments__assignment_id__printable_get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintableAssignment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Printable Assignment",
        "tags": [
          "assignments"
        ]
      }
    },
    "/v1/assignments/{assignment_id}/start": {
      "post": {
        "operationId": "start_assignment_v1_assignments__assignment_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentWork"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Assignment",
        "tags": [
          "assignments"
        ]
      }
    },
    "/v1/attempts/{attempt_id}/correction": {
      "post": {
        "operationId": "create_correction_v1_attempts__attempt_id__correction_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentWork"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Correction",
        "tags": [
          "attempts"
        ]
      }
    },
    "/v1/attempts/{attempt_id}/responses/{question_id}": {
      "put": {
        "operationId": "save_response_v1_attempts__attempt_id__responses__question_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Response",
        "tags": [
          "attempts"
        ]
      }
    },
    "/v1/attempts/{attempt_id}/results": {
      "get": {
        "operationId": "get_results_v1_attempts__attempt_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Results",
        "tags": [
          "attempts"
        ]
      }
    },
    "/v1/attempts/{attempt_id}/submit": {
      "post": {
        "operationId": "submit_attempt_v1_attempts__attempt_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Attempt",
        "tags": [
          "attempts"
        ]
      }
    },
    "/v1/attempts/{attempt_id}/work": {
      "get": {
        "operationId": "get_attempt_work_v1_attempts__attempt_id__work_get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentWork"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Attempt Work",
        "tags": [
          "attempts"
        ]
      }
    },
...
    "/v1/question-sets/{question_set_id}/assignments": {
      "post": {
        "operationId": "assign_question_set_v1_question_sets__question_set_id__assignments_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Question Set",
        "tags": [
          "question-sets"
        ]
      }
    },
...
  }
}